+++ src/pages/project/project_wizard/CreateProject.tsx
@@ -1,12 +1,11 @@
-import { Box } from '@mui/material';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import { useFetchCurrentInstitution } from '../../../api/hooks/useFetchCurrentInstitution';
 import { PageHeader } from '../../../components/PageHeader';
-import { StyledPageContent } from '../../../components/styled/Wrappers';
+import { StyledPageContent, WizardStartPageWrapper } from '../../../components/styled/Wrappers';
 import { RootState } from '../../../redux/store';
-import { CristinProject, emptyProject, SaveCristinProject } from '../../../types/project.types';
+import { emptyProject, SaveCristinProject } from '../../../types/project.types';
 import { EmptyProjectForm } from './EmptyProjectForm';
 import { NFRProject } from './NFRProject';
 import { ProjectForm } from './ProjectForm';
@@ -15,25 +14,18 @@ const CreateProject = () => {
   const { t } = useTranslation();
   const user = useSelector((store: RootState) => store.user);
   const topOrgCristinId = user?.topOrgCristinId ?? '';
+  const currentInstitutionQuery = useFetchCurrentInstitution(topOrgCristinId);
   const [newProject, setNewProject] = useState<SaveCristinProject>({ ...emptyProject });
   const [showProjectForm, setShowProjectForm] = useState<boolean>(false);
-
-  const currentInstitutionQuery = useFetchCurrentInstitution(topOrgCristinId);
 
   return (
     <StyledPageContent>
       {showProjectForm ? (
-        <ProjectForm project={newProject as CristinProject} />
+        <ProjectForm project={newProject} />
       ) : (
         <>
           <PageHeader>{t('project.create_project')}</PageHeader>
-          <Box
-            sx={{
-              maxWidth: '55rem',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: '1.25rem',
-            }}>
+          <WizardStartPageWrapper>
             <NFRProject
               newProject={newProject}
               setNewProject={setNewProject}
@@ -45,7 +37,7 @@ const CreateProject = () => {
               setNewProject={setNewProject}
               setShowProjectForm={setShowProjectForm}
             />
-          </Box>
+          </WizardStartPageWrapper>
         </>
       )}
     </StyledPageContent>
